+++ include/cpu/inorder/cpu.h
@@ -71,7 +71,7 @@ private:
     
     // CPU 状态
     std::array<uint64_t, NUM_REGISTERS> registers_;
-    std::array<uint32_t, NUM_FP_REGISTERS> fp_registers_;
+    std::array<uint64_t, NUM_FP_REGISTERS> fp_registers_;
     std::array<uint64_t, NUM_CSR_REGISTERS> csr_registers_;
     uint64_t pc_;                   // 程序计数器
     bool halted_;                   // 停机标志
